Add the chsmStart() method to the CHsm class.
CHsm's constructor tries to do the initial transition of the derived
state machine.  This causes problems as the base class constructor
runs and finishes before the derived class constructor, so parts of
the derived class may not exist during the base constructor.  If some
of the actions that run during the initial transition depend on those
parts, we get runtime errors, commonly segfaults.

So put this initial transition into a chsmStart() method, and require
that derived classes call that before doing any other event activity.
This requirement is enforced with an assertion.

Signed-off-by: Russell Steicke <russells@adelie.cx>