Change the mechanism used to find out a state's parent.
If a state has a parent state, then in response to the CTHE_PARENT
event, or any other event that it does not handle, it needs to set
_parentState and return CTH_PARENT.  (This can be done by calling
cth_parent(state) and returning its return value.)  If it does not
have a parent state, which can only be true for the single top state
of a HSM, then it returns CTH_HANDLED.  (This can be done by calling
cth_handled() and returning its return value.)

As a consequence of this, the top state does not have to be specified
in the constructor call.  It is discovered later, when we walk up the
hierarchy.  The initial state can even be the top state, and it can
transition to other states to begin whatever processing is required of
the HSM.

Prior to this change, the top state would never be called at all, and
its only purpose was to have its address taken (as a pointer to
member) and have that address compared to other parent states when we
were walking up a HSM hierarchy.  Now the top state can be called, and
can possibly perform some actions in response to other events.

The template class still contains a topState() function which does
nothing except return CTH_HANDLED.  This topState() can be used as the
top state for any HSM.

CTH_I_AM_THE_TOP_STATE is an alias for CTH_HANDLED.  It is intended
that the top state return this value as documentation, making it
easier to identify the top state.

There is another consequence of this change.  A HSM can now transition
to its top state, which was forbidden before, and that top state can
now handle events and do other transitions.  Because of this, if a HSM
uses the default template topState() as its top state, it should never
transition there.  If it does, all subsequent events will be ignored.
Effectively, the HSM will be a zombie that takes everything in but
does nothing.

Signed-off-by: Russell Steicke <russells@adelie.cx>