Rename test shell scripts and directories.
The .d extension was being used for two purposes: dependency files and
test data directories.  These conflicted in that git is required to
ignore the first but not the second, and adding files inside those
directories is more hassle than it should be.  Also, new files in
those directories are not noticed and reported as untracked by git.

So rename all the shell scripts from tXX to tXX.sh and the directories
from tXX.d to tXX.

Signed-off-by: Russell Steicke <russells@adelie.cx>